+++ src/index.js
@@ -1,14 +1,10 @@
 #!/usr/bin/env node
 
-var version = "1.3.2";
+var version = "1.3.3";
 var methods = require("./methods.js");
 var messages = require("./messages.js");
 var processes = require("./process.js");
 var command = process.argv[2];
-
-// Display intro message
-console.log("\n");
-messages.outputMessage("Cordova Clean", true);
 
 // Check for clean config file
 var options = methods.getCleanConfig("cordova-clean.json");
@@ -35,16 +31,6 @@ if (options.noAndroid) {
     options.skipList.push("android");
 }
 
-// Display which cli options were included
-messages.outputOptions(options);
-
-// Extract the config.xml content
-var config = methods.loadConfig();
-if (!config) {
-    messages.outputMessage("Done", true);
-    return;
-}
-
 // Process version command before others, that way we don't log intro / outro messages
 switch (command) {
     case "--version":
@@ -57,23 +43,35 @@ switch (command) {
         return;
 }
 
+// Display intro message
+console.log("\n");
+messages.outputMessage("Cordova Clean", true);
+
+// Make sure we have all requirements
+if (!methods.checkRequirements()) {
+    return;
+}
+
+// Display which cli options were included
+messages.outputOptions(options);
+
 switch (command) {
     case "clean":
-        processes.removePlatforms(config, options);
-        processes.removePlugins(config, options);
-        processes.addPlugins(config, options);
-        processes.addPlatforms(config, options);
+        processes.removePlatforms(options);
+        processes.removePlugins(options);
+        processes.addPlugins(options);
+        processes.addPlatforms(options);
         break;
     case "plugins":
-        processes.removePlugins(config, options);
-        processes.addPlugins(config, options);
+        processes.removePlugins(options);
+        processes.addPlugins(options);
         break;
     case "platforms":
-        processes.removePlatforms(config, options);
-        processes.addPlatforms(config, options);
+        processes.removePlatforms(options);
+        processes.addPlatforms(options);
         break;
     case "sync":
-        processes.sync(config, options);
+        processes.sync(options);
         break;
     default:
         console.log(messages.helpMessage);